requirements.txt
cycler==0.10.0
decorator==4.4.2
et-xmlfile==1.0.1
imageio==2.9.0
jdcal==1.4.1
kiwisolver==1.3.1
matplotlib==3.3.3
networkx==2.5
numpy==1.19.4
opencv-python==4.4.0.46
openpyxl==3.0.5
Pillow==8.0.1
pyparsing==2.4.7
pytesseract==0.3.6
python-dateutil==2.8.1
PyWavelets==1.1.1
scikit-image==0.17.2
scipy==1.5.4
six==1.15.0
tifffile==2020.10.1
